requirements.txt
# Core dependencies for AIREX project
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0

# Testing dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Development dependencies (optional)
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0
